script/requirements.txt
psycopg2-binary
paho-mqtt==1.6.1
pytz
requests
python-dotenv
